JWK creation and validation improvements.
LittleJWTBuilder allows JWKValidator to be specified.
The default JWKValidator can be specified.
LittleJWT is created with JWK validator.
A random JWK is generated if JWK validation fails (by default).
The JWK is passed when JWKValidator is invoked, not constructed.
A fallback for when JWK validation fails can be specified.
JWKValidator uses AlgorithmBuilder to validate JWK algorithm.